requirements.txt
openai>=1.40.0
pandas
python-dotenv
rich
tenacity
tqdm